Added cache annotations and expiry

+++ src/main/java/com/example/demo/dao/StudentRepo.java
@@ -1,8 +1,12 @@
 package com.example.demo.dao;
 
 import com.example.demo.model.Student;
+import org.springframework.cache.annotation.Cacheable;
 import org.springframework.data.repository.CrudRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface StudentRepo extends CrudRepository<Student, Long> {}
+public interface StudentRepo extends CrudRepository<Student, Integer> {
+
+
+}
